Fix most warnings (mainly tests)

+++ test/test_dict.c
@@ -21,7 +21,7 @@ int dict_test_teardown(void **state) {
 int verify_dict_size(dict * restrict d) {
   size_t bucket_sum = 0, i = 0;
 
-  while (i < 1 << d->exponent) {
+  while (i < (size_t) 1 << d->exponent) {
     if (d->buckets[i].flags & 0x1) {
       bucket_sum += ((list *) d->buckets[i].data)->size;
     } else if (d->buckets[i].data) {
